requirements.txt
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
coverage==7.6.9
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
nest-asyncio==1.6.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.25.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-timeout==2.3.1
python-dotenv==1.0.0
python-telegram-bot==21.9
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
TgCrypto==1.2.5
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
